Added common module to maintain common code as JAR and add dependency in all submodules
