+++ backend-common/src/main/java/com/online/ems/CommonEMSApplication.java
@@ -1,0 +1,8 @@
+package com.online.ems;
+
+public class CommonEMSApplication {
+
+	public static void main(String[] args) {
+		
+	}
+}
