+++ backend-common/src/main/java/com/online/ems/bean/EMSBean.java
@@ -1,0 +1,22 @@
+package com.online.ems.bean;
+
+public class EMSBean {
+	private String title;
+	private String environment;
+
+	public String getTitle() {
+		return title;
+	}
+
+	public void setTitle(String title) {
+		this.title = title;
+	}
+
+	public String getEnvironment() {
+		return environment;
+	}
+
+	public void setEnvironment(String environment) {
+		this.environment = environment;
+	}
+}
