--- backend-employee-service/src/main/java/com/online/ems/bean/DepartmentsBean.java
@@ -1,23 +1,0 @@
-package com.online.ems.bean;
-
-public class DepartmentsBean {
-	
-	private String deptNo;
-	private String deptName;
-
-	public String getDeptNo() {
-		return deptNo;
-	}
-
-	public void setDeptNo(String deptNo) {
-		this.deptNo = deptNo;
-	}
-
-	public String getDeptName() {
-		return deptName;
-	}
-
-	public void setDeptName(String deptName) {
-		this.deptName = deptName;
-	}
-}
