+++ backend-salaries-service/pom.xml
@@ -18,6 +18,12 @@
 
 	<dependencies>
 		<dependency>
+			<groupId>employee-management-system</groupId>
+			<artifactId>backend-common</artifactId>
+			<version>1.0.0</version>
+		</dependency>
+	
+		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-web</artifactId>
 		</dependency>
